requirements.txt
Flask==2.3.2
tensorflow==2.15.0
Pillow==10.0.1
numpy==1.26.0
reportlab==4.4.4
